neuro-career-be/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# AI and ML
google-generativeai==0.3.2
elevenlabs==0.2.26
assemblyai==0.20.0

# Audio processing
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3

# Web and API
httpx==0.25.2
requests==2.31.0

# Optional: for CORS and additional features
python-multipart==0.0.6